fix(sonar): C++20 mechanical cleanup batch (S6164/S6177/S6004/S3608)
Replace magic ln2/pi with std::numbers, add function-scope using enum,
if-init-statements, and explicit lambda captures. Also fix quick singleton
issues: agent MotilityState declaration, motility empty compute comment,
string_view parse_bool, merged oxygen if, fix_registry structured binding
and try_emplace, spatial_hash_gpu access layout, CellKey =default,
LineageEvent enum class, contains(), redundant cast removal.

+++ tests/test_hdf5_roundtrip.cpp
@@ -312,7 +312,7 @@ void validate_step(hid_t file,
                    Real expected_time) {
   validate_step_schema(file, step);
   validate_step_metadata(file, step, expected_step, expected_time,
-                         static_cast<Int>(sim.global_agent_count()));
+                         sim.global_agent_count());
   validate_step_agents_match_sim(file, step, sim);
   validate_step_genome(file, step, sim);
 }
